# Performance improvements to the Parser