--- a/src/Combine.elm
+++ b/src/Combine.elm
@@ -31,7 +31,6 @@ module Combine exposing
     , runParser
     , sepBy
     , sepBy1
-    , sepBy1Core
     , sepBy1WithState
     , sepBy1WithoutReverse
     , succeed
@@ -238,26 +237,6 @@ manyIgnore (Parser p) =
             Core.loop state helper
 
 
-manyCore : Core.Parser a -> Core.Parser (List a)
-manyCore p =
-    manyWithoutReverseCore [] p
-        |> Core.map List.reverse
-
-
-manyWithoutReverseCore : List a -> Core.Parser a -> Core.Parser (List a)
-manyWithoutReverseCore initList p =
-    let
-        helper : List a -> Core.Parser (Core.Step (List a) (List a))
-        helper items =
-            Core.oneOf
-                [ p
-                    |> Core.map (\item -> Core.Loop (item :: items))
-                , Core.succeed (Core.Done items)
-                ]
-    in
-    Core.loop initList helper
-
-
 many1Ignore : Parser state () -> Parser state ()
 many1Ignore p =
     p
@@ -359,17 +338,6 @@ sepBy1WithState sep p =
     succeed cons
         |> keep p
         |> keep (many (sep |> continueWith p))
-
-
-sepBy1Core : String -> Core.Parser a -> Core.Parser (List a)
-sepBy1Core sep p =
-    Core.succeed cons
-        |= p
-        |= manyCore
-            (Core.succeed identity
-                |. Core.symbol sep
-                |= p
-            )
 
 
 {-| Same as [`sepBy1`](#sepBy1), except that it doesn't reverse the list.