--- a/src/Combine/Char.elm
+++ b/src/Combine/Char.elm
@@ -1,31 +1,17 @@
-module Combine.Char exposing (anyChar, char)
+module Combine.Char exposing (anyChar)
 
 import Parser as Core
 
 
-char : Char -> Core.Parser Char
-char c =
-    satisfy
-        (\c_ -> c_ == c)
-        ("expected '" ++ String.fromChar c ++ "'")
-
-
 anyChar : Core.Parser Char
 anyChar =
-    satisfy
-        (always True)
-        "expected any character"
-
-
-satisfy : (Char -> Bool) -> String -> Core.Parser Char
-satisfy pred problem =
-    Core.chompIf pred
+    Core.chompIf (always True)
         |> Core.getChompedString
         |> Core.andThen
             (\s ->
                 case String.toList s of
                     [] ->
-                        Core.problem problem
+                        Core.problem "expected any character"
 
                     c :: _ ->
                         Core.succeed c