… review/suppressed/NoImportingEverything.json
… src/Elm/Parser/Expression.elm
… src/Elm/Parser/File.elm
… src/Elm/Parser/Imports.elm
… src/Elm/Parser/Layout.elm
… src/Elm/Parser/Modules.elm
… src/Elm/Parser/Node.elm
… src/Elm/Parser/Numbers.elm
… src/Elm/Parser/Patterns.elm
… src/Elm/Parser/Tokens.elm
… src/Elm/Parser/TypeAnnotation.elm
… src/Elm/Parser/Typings.elm
… src/Elm/Parser/Whitespace.elm
… src/List/Extra.elm
… src/Parser/Extra.elm
… src/Pratt.elm
… tests/Elm/Parser/BaseTest.elm
… tests/Elm/Parser/CombineTestUtil.elm
… tests/Elm/Parser/NumbersTests.elm
… tests/Elm/Parser/TestUtil.elm
… tests/Elm/Parser/TokenTests.elm
… tests/Elm/Parser/WhitespaceTests.elm