Follow up to #222

(Metrics are for my machine, not directly comparable to @jiegillet's results)

Before pratt-parser: 73400.5 ms to do 100 runs => 734.005 ms/run (weird, @jiegillet's results for this was much slower?)
Before: 146231.60000002384 ms to do 100 runs => 1462.3160000002383 ms/run
After: 101167.20000004768 ms to do 100 runs => 1011.6720000004768 ms/run

There are plenty of optimizations done here:
- Improve the arity of functions (see https://blogg.bekk.no/how-elm-functions-work-71cab7426a2f), which leads to some odd code in applicatives
- Trying to avoid using `Combine` as much as possible, and using the core parser as much as possible instead, wrapping as late as possible.
	- `Combine` has a lot of overhead, so sticking to using core parser is much better.
- Try to avoid unnecessary collection of data for things that get ignored (this is now enforced when using `Combine.ignore` and friends)