@jfmengels 2024-07-16
Alright, I think I'm done with this batch. I know I've introduced some complexity (I'm mostly thinking all the core->combine and combine->core functions). If anyone works with this, feel free to change or refactor it. This was an incremental change

Here are my most recent benchmarks:
- Master: 970 ms/run
- Before: 1543 ms/run
- After: 1065 ms/run

In other words, we're pretty close to the speed from before, although I'm surprised we're not faster. @jiegillet could you try running these benchmarks on your device?

@jiegillet If you feel like reviewing, please do, but it's also quite alright to say you don't want to look at...  *checks notes* ... 125 commits :sweat_smile: