@jiegillet 2024-07-17
Very cool!
I didn't check commit-by-commit, but it overall looks good. 
Some of the changes did make the code more complex, but it's hard to avoid for an optimization PR.

as for benchmarks on my machine:
master: 1374 ms/run
before: 1579 ms/run
after: 1262 ms/run

very impressive! Although these numbers are pretty unreliable, this is definitely on par with master!