@jfmengels 2024-07-17
@jiegillet do you mean to say that for you this branch (not including the other PRs up until #225) is already faster than master? That's great if it is, though not sure how we get such different benchmarks.