@jiegillet 2024-07-17
Yes, that's what I got.
I guess 100 runs is really not reliable, one bad run makes the average much longer. The time to run with the default settings is like an hour and a half after all :)
I wouldn't worry about it, we both see that it's faster, and the code is still readable, it's great that you found even more ways to optimize the Pratt parser.